game uno adapter class and fav icon update
